Resolve conflicts after ~39k upstream commits:
- fix.h / comm_kokkos.{cpp,h} / atom_kokkos.cpp: take develop. It already
  provides reverse_comm_device, per-fix host/device comm dispatch, and the
  per-fix sort_device host fallback that the PR had hacked in, so the PR's
  per-timestep and *_classic workarounds are superseded and dropped.
- math_extra_kokkos.h: take develop's header (rigid-body-specific functions
  are added back in a follow-up); drop the PR's math_extra_kokkos.cpp.
- RIGID/fix_rigid_small.h: keep develop's bigint nbody and resample_momenta;
  keep the PR's virtual qualifiers and public struct Body (needed by the
  Kokkos subclass / device lambdas).
- Install.sh: keep develop's CMake-only structure; add rigid/small kk entries.
- Remove a stray debug printf block left in unpack_reverse_comm.

+++ src/BODY/pair_body_rounded_polygon.cpp
@@ -763,8 +763,8 @@ int PairBodyRoundedPolygon::vertex_against_edge(int i, int j,
 
         // vertex i interacts with a vertex of the edge, but does not contact
 
-        if (mode == VERTEXI) p2vertex = edge[jefirst+nj][0];
-        else if (mode == VERTEXJ) p2vertex = edge[jefirst+nj][1];
+        if (mode == VERTEXI) p2vertex = (int)edge[jefirst+nj][0];
+        else if (mode == VERTEXJ) p2vertex = (int)edge[jefirst+nj][1];
 
         // double xj[3];
         // p2.body2space(p2vertex, xj);
