+++ src/CG-SPICA/pair_lj_spica_coul_long.cpp
@@ -185,7 +185,7 @@ template <int EVFLAG, int EFLAG, int NEWTON_PAIR> void PairLJSPICACoulLong::eval
             rsq_lookup.f = rsq;
             itable = rsq_lookup.i & ncoulmask;
             itable >>= ncoulshiftbits;
-            fraction = (rsq_lookup.f - rtable[itable]) * drtable[itable];
+            fraction = ((double) rsq_lookup.f - rtable[itable]) * drtable[itable];
             table = ftable[itable] + fraction * dftable[itable];
             forcecoul = qtmp * q[j] * table;
             if (EFLAG) ecoul = qtmp * q[j] * (etable[itable] + fraction * detable[itable]);
@@ -579,7 +579,7 @@ double PairLJSPICACoulLong::single(int i, int j, int itype, int jtype, double rs
       rsq_lookup_single.f = rsq;
       itable = rsq_lookup_single.i & ncoulmask;
       itable >>= ncoulshiftbits;
-      fraction = (rsq_lookup_single.f - rtable[itable]) * drtable[itable];
+      fraction = ((double) rsq_lookup_single.f - rtable[itable]) * drtable[itable];
       table = ftable[itable] + fraction * dftable[itable];
       forcecoul = atom->q[i] * atom->q[j] * table;
       table = etable[itable] + fraction * detable[itable];
