+++ src/CORESHELL/pair_coul_long_cs.cpp
@@ -133,7 +133,7 @@ void PairCoulLongCS::compute(int eflag, int vflag)
           rsq_lookup.f = rsq;
           itable = rsq_lookup.i & ncoulmask;
           itable >>= ncoulshiftbits;
-          fraction = (rsq_lookup.f - rtable[itable]) * drtable[itable];
+          fraction = ((double) rsq_lookup.f - rtable[itable]) * drtable[itable];
           table = ftable[itable] + fraction*dftable[itable];
           forcecoul = scale[itype][jtype] * qtmp*q[j] * table;
           if (factor_coul < 1.0) {
