+++ src/DIFFRACTION/fix_saed_vtk.cpp
@@ -172,7 +172,7 @@ FixSAEDVTK::FixSAEDVTK(LAMMPS *lmp, int narg, char **arg) :
     if ((Zone[0] == 0) && (Zone[1] == 0) && (Zone[2] == 0)) {
       for (int i=0; i<3; i++) {
         dK[i] = prd_inv[i]*c[i];
-        Knmax[i] = ceil(Kmax / dK[i]);
+        Knmax[i] = (int) ceil(Kmax / dK[i]);
         Knmin[i] = -Knmax[i];
       }
     } else {
@@ -187,7 +187,7 @@ FixSAEDVTK::FixSAEDVTK(LAMMPS *lmp, int narg, char **arg) :
       int Ksearch[3];
       for (int i=0; i<3; i++) {
         dK[i] = prd_inv[i]*c[i];
-        Ksearch[i] = ceil(Kmax / dK[i]);
+        Ksearch[i] = (int) ceil(Kmax / dK[i]);
       }
 
       for (int k = -Ksearch[2]; k <= Ksearch[2]; k++) {
