+++ src/ML-SNAP/compute_gaussian_grid_local.cpp
@@ -16,7 +16,6 @@
 #include "atom.h"
 #include "comm.h"
 #include "error.h"
-#include "force.h"
 #include "math_const.h"
 #include "math_special.h"
 #include "memory.h"
@@ -24,7 +23,6 @@
 #include "update.h"
 
 #include <cmath>
-#include <cstring>
 
 using namespace LAMMPS_NS;
 using MathConst::MY_2PI;
@@ -160,7 +158,7 @@ void ComputeGaussianGridLocal::compute_local()
 double ComputeGaussianGridLocal::memory_usage()
 {
   int n = atom->ntypes + 1;
-  int nbytes = (double) n * sizeof(int);    // map
+  double nbytes = (double) n * sizeof(int);    // map
 
   return nbytes;
 }
